backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.12
pandas==2.2.3
numpy==2.1.1
openai==1.51.0
python-dotenv==1.0.1
aiofiles==24.1.0
